Make rds-hq sensor optional and move script to nvidia/cosmos directories
