+++ PythonAPI/examples/nvidia/cosmos/client/cosmos_aov.yaml
@@ -1,7 +1,3 @@
-# SPDX-FileCopyrightText: © 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-#
-# SPDX-License-Identifier: MIT
-
 - sensor: rgb
   attributes:
     image_size_x: 1920
@@ -67,6 +63,20 @@
     location:
       x: 0
       y: 0
+      z: 1.8
+    rotation:
+      pitch: 8
+      yaw: 0
+      roll: 0
+- sensor: rds_hq
+  attributes:
+    image_size_x: 1920
+    image_size_y: 1080
+    fov: 110
+  transform:
+    location:
+      x: 0.0
+      y: 0.0
       z: 1.8
     rotation:
       pitch: 8
